+++ kernel/capability.c
@@ -19,7 +19,7 @@ spinlock_t task_capability_lock;
 
 kernel_cap_t cap_bset = CAP_FULL_SET;
 
-asmlinkage int sys_capget(cap_user_header_t header, cap_user_data_t dataptr)
+asmlinkage long sys_capget(cap_user_header_t header, cap_user_data_t dataptr)
 {
      int error, pid;
      __u32 version;
@@ -126,7 +126,7 @@ static void cap_set_all(kernel_cap_t *effective,
  * E: must be set to a subset of (new target) Permitted
  */
 
-asmlinkage int sys_capset(cap_user_header_t header, const cap_user_data_t data)
+asmlinkage long sys_capset(cap_user_header_t header, const cap_user_data_t data)
 {
      kernel_cap_t inheritable, permitted, effective;
      __u32 version;
